wp-now: Require node v18 or higher to run (#92)
In this patch we're calling a wrapper script that checks the version of
Node before executing code from the project.

Previously, the version check occurred within the same running Node
process as the project code, and so on unsupported Node versions it
could crash while reading in the initial code to run.

By moving the check out of process we can ensure that the minimum syntax
and library support is available before we attempt to run any code.

cc: @danielbachhuber 

I haven't yet determined if I can run this with the wrapper script, but
I independently tested the wrapper script and I think it's reliable
enough. As for your questions about about command-line arguments I
believe it's taken care of by using the `"$@"` in quotes like that; it's
a special form added exactly to fix broken string interpolation issues
with arguments.