Add "playground" keyword to Playground Block. (#148)
This is a minor tweak, but I bet that many people will search for
"Playground" or type `/playground` when trying to insert the Playground
Block. The addition of the `keyword` property in `block.json` makes this
possible.

Currently, you have to search for "WordPress..." for the block to show
up.