Wait for wordpress to fully unzip before moving it to ~/.wp-now
There is a race condition in the code that downloads the latest
WordPress code, unzips it, and places it in `~/.wp-now`.

The error happens in the final step which moves the files. My
understanding is that Macs are more forgiving than Windows when it comes
to renaming/moving files, and it seems that's the issue here. The
current code doesn't wait for the unzipping to finish before moving the
files. And that causes a permission error on Windows.

This diff waits for each entry in the zip file to finish being written
before progressing on to moving the zip contents. It waits for the
'close' event. Usually it's sufficient to wait for the 'finish' event
before proceeding. However we need to wait for the underlying file
object to be released, and that's the 'close' event.

It also Makes sure we "drain" each entry in the zip file. The docs for
`unzipper` make a point that each entry in the archive needs to be fully
consumed. I'm not sure how important this actually is for directory
entries, but I've added it for good measure.