Playground block: Import the client library from playground.wordpres.net (#174)
Importing the library from an npm package requires updating the plugin
each time a new version is released. This ensures the block always uses
the latest version.

## Testing instructions

Run the block locally with `nx dev wordpress-playground-block` and
confirm it works as expected. Then build it with `nx build
wordpress-playground-block`, install on a WordPress site, and confirm
the built version also works.