WordPress Playground Block: Load the Edit component asynchronously
Loads the edit component asynchronously. This may take a while,
and it seem to involve a few top-level async resolutions which
delay the `registerBlockType()` call below when imported directly.

As a result, the editor sometimes displays a "Block is not available" error.

The dynamic import introduced in this PR is async and allows the registerBlockType()
call to be synchronous, which allows the block to reliably load.

 ## Testing instructions

Create a page with the Playground block in is. Save. Reload it 10 times.
Confirm the block loaded correctly each time.