Vscode extension: update WordPress versions (#185)
Updates the WordPress versions displayed in the UI. The latest available
version is now 6.4.

## Testing instructions

Install the extensions using the instructions from its README and
confirm the latest available WP version is now 6.4.

## Follow-up work

Retrieve the available versions from the public API endpoint when a
network connection is available. Otherwise, make a note about the
network connection being down and display the versions present in a
local cache.