Fix side-by-side code editor to be full height within container (#191)
This PR fixes an issue with code editor height. Before this PR, there
were some cases where the code editor left some bottom space between it
and the Run bar blank and unoccupied.

Before:

![before-height-fix](https://github.com/WordPress/playground-tools/assets/530877/52577f2a-8f47-439a-919e-852fb6c0497c)

After:

![after-height-fix](https://github.com/WordPress/playground-tools/assets/530877/8674fc71-c5bb-4925-a9b9-555a84153384)

## Why?

It can be aesthetically displeasing and sometimes confusing to have
open, purposeless space within the block.

## How?

This PR changes max-height from
`min(450px, 80vh)`
to
`100%`.

It is possible this is the wrong solution, but I do not currently see a
reason we wouldn't want the editor pane to take 100% of the available
size.

## Testing Instructions

Check out this branch, follow the [Contributing
instructions](https://github.com/WordPress/playground-tools/tree/trunk/packages/wordpress-playground-block#contributing)
to build and start a dev version of WordPress, create a post containing
the block, and observe the block rendering in the editor and on the
front end.