wordpress-playground-block: Require Activation Before Loading Playground Iframe (#183)
This PR updates the wordpress-playground-block to only load Playground
after an Activate button has been clicked.

## Why?

IFrames can create difficulty for those using Screen Readers. See:
[#](https://github.com/WordPress/Learn/pull/2264#issuecomment-1971503159),
[#](https://github.com/WordPress/Learn/pull/2264#issuecomment-1979277253)

Requiring user action to load Playground will help avoid distracting
screen readers with auto-loading iframes. As a bonus, it would also stop
the block from automatically consuming network bandwidth to download
Playground assets, even if Playground will not be used.

## How?

<!-- How is your PR addressing the issue at hand? What are the
implementation details? -->
This button must be clicked before the code editor and Playground load,
so the iframe no longer loads immediately.

## Questions

Would it be better to just gate the iframe with an Activate button?
Perhaps that doesn't make sense because of the expected interaction with
the code editor, but I think it _might_ look better than a single
activation button in the middle of a large block like this one.