Add express to dependencies; Update follow-redirects to newest version (#225)
This is the second part of solving:

- #224 

The first part is WordPress/wordpress-playground#1218. (php-wasm/node:
Update express to newest version, and move it to devDependencies)

## Why?

The first part of the solution is in the Playground project, which
updates `express` and moves it to `devDependencies` since it's only used
for tests in that project.

As a result, `express` will no longer be installed when `@php-wasm/node`
is installed. So `wp-now` needs to declare `express` as a dependency.

## How?

- Add `express` to `dependencies`
- Update `follow-redirects` to newest `1.15.6`

These together with WordPress/wordpress-playground#1218 will resolve the
original issue #224, which should eliminate all npm audit warnings.

## Testing Instructions

Currently `npm run test` fails with an unrelated issue, as can be seen
in another PR's CI test run.
https://github.com/WordPress/playground-tools/actions/runs/8616282863/job/23614052137?pr=223#step:4:49
("RuntimeError: memory access out of bounds") That seems have been
caused by commit
[133029c](https://github.com/WordPress/playground-tools/commit/133029c236cffa5cee97d599354cb33393b883b6).