New tool: Edit visually Browser Extension (#285)
## What?

Adds an experimental browser extension that allows you to edit GitHub
issues and other text formats using WordPress blocks in Playground.

![CleanShot 2024-06-02 at 17 26
11@2x](https://github.com/WordPress/playground-tools/assets/205419/a8eaafaa-5890-4134-aa08-5e4dda2d89b5)

To try it, go to the extensions page in your web browser, load the
extension from the `packages/editor-browser-extension` directory, and
then navigate to any GitHub issue. Edit it, and you should see a new
"Edit in Playground" button in the corner of the editor. Click it to
edit the issue in Playground.

There's a bunch of cleaning up that could be done here – I know.

cc @dmsnell @bgrgicak @brandonpayton