Playground block: a11y: Speak notice dismissal (#348)
## What and why?

This PR adds a spoken notice for those using assistive technologies when
the exit-the-editor-with-keyboard notice is dismissed.

Related to #343

## How?

The Playground block is wrapped using `withSpokenMessages()` from
`@wordpress/components`. The spoken message is "polite" and does not
interrupt the screen reader.

Prior to this PR, the editor itself was focused when the
exit-with-keyboard notice was dismissed, but this was not good for two
reasons:
1. It triggers reading editor content which is potentially very long.
2. When the editor text is being read, the user will have to wait a long
time for the "Dismissed notice" message to be spoken.

This PR now focuses the previous item in the tab order (the download
button) when the exit-with-keyboard notice is dismissed.

## Testing Instructions

- Enable a screen reader
- Run npx nx dev wordpress-playground-block
- View a post that has the Playground block
- Move focus to the exit-with-keyboard notice at the top of the editor
- Click the notice to dismiss it
- Confirm that "Notice dismissed" is spoken after the notice is
dismissed. This may happen after the download button text is read.

To repeat this, open the dev tools and delete the local storage key
"playground-block-dismiss-exit-editor-tip" and refresh the page.